refactor: move to /chat/ folders
