+++ src/main/java/org/runejs/client/message/handler/rs435/chat/ForcedPrivateMessageHandler.java
@@ -1,9 +1,9 @@
-package org.runejs.client.message.handler.rs435;
+package org.runejs.client.message.handler.rs435.chat;
 
 import org.runejs.client.frame.ChatBox;
 import org.runejs.client.media.renderable.actor.Player;
 import org.runejs.client.message.handler.MessageHandler;
-import org.runejs.client.message.inbound.ForcedPrivateMessageInboundMessage;
+import org.runejs.client.message.inbound.chat.ForcedPrivateMessageInboundMessage;
 
 /**
  * A {@link MessageHandler} that handles {@link ForcedPrivateMessageInboundMessage}s.
