+++ src/main/java/org/runejs/client/message/handler/rs435/chat/ReceivePrivateMessageHandler.java
@@ -1,11 +1,11 @@
-package org.runejs.client.message.handler.rs435;
+package org.runejs.client.message.handler.rs435.chat;
 
 import org.runejs.client.MovedStatics;
 import org.runejs.client.frame.ChatBox;
 import org.runejs.client.language.Native;
 import org.runejs.client.media.renderable.actor.Player;
 import org.runejs.client.message.handler.MessageHandler;
-import org.runejs.client.message.inbound.ReceivePrivateMessageInboundMessage;
+import org.runejs.client.message.inbound.chat.ReceivePrivateMessageInboundMessage;
 import org.runejs.client.util.TextUtils;
 
 /**
