+++ src/main/java/org/runejs/client/message/handler/rs435/chat/SetFriendsListStatusMessageHandler.java
@@ -1,9 +1,9 @@
-package org.runejs.client.message.handler.rs435;
+package org.runejs.client.message.handler.rs435.chat;
 
 import org.runejs.client.cache.media.gameInterface.GameInterface;
 import org.runejs.client.media.renderable.actor.Player;
 import org.runejs.client.message.handler.MessageHandler;
-import org.runejs.client.message.inbound.SetFriendsListStatusInboundMessage;
+import org.runejs.client.message.inbound.chat.SetFriendsListStatusInboundMessage;
 
 /**
  * A {@link MessageHandler} that handles {@link SetFriendsListStatusInboundMessage}s.
