+++ src/main/java/org/runejs/client/net/codec/runejs435/RuneJS435PacketCodec.java
@@ -2,7 +2,7 @@ package org.runejs.client.net.codec.runejs435;
 
 import org.runejs.client.net.PacketType;
 import org.runejs.client.net.codec.MessagePacketCodec;
-import org.runejs.client.net.codec.runejs435.decoder.*;
+import org.runejs.client.net.codec.runejs435.decoder.chat.*;
 
 /**
  * A {@link MessagePacketCodec} for the RuneJS customised 435 protocol.
